Split spotify_handlers.rs into modules by concern
The 652-line file mixed three unrelated layers: artist lookup, the
OAuth login/callback/status flow, playlist HTTP handlers, and raw
sqlx DB access for sessions/accounts/playlists — all sharing one
namespace with no clear ownership.

Split into spotify_handlers/{artist,auth,playlists,db}.rs along
those boundaries. db.rs holds all persistence and is pub(super), so
only sibling modules within spotify_handlers can reach it directly.
mod.rs glob re-exports the public items, so every existing
`spotify::spotify_handlers::*` call site (all in lib.rs route
wiring) needed zero changes. No behavior change — same routes, same
queries, same tests passing.
